// src/apollo/types/types.ts
export enum EProjectSocialMediaType {
  FACEBOOK = "FACEBOOK",
  X = "X",
  INSTAGRAM = "INSTAGRAM",
  YOUTUBE = "YOUTUBE",
  LINKEDIN = "LINKEDIN",
  REDDIT = "REDDIT",
  DISCORD = "DISCORD",
  FARCASTER = "FARCASTER",
  LENS = "LENS",
  WEBSITE = "WEBSITE",
  TELEGRAM = "TELEGRAM",
  GITHUB = "GITHUB",
}

export interface IProjectSocialMedia {
  type: EProjectSocialMediaType;
  link: string;
}